backend/requirements.txt
asgiref==3.11.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.2.9
django-allauth==65.13.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gunicorn==23.0.0
idna==3.11
oauthlib==3.3.1
packaging==25.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.5
urllib3==2.6.2
Pillow==10.4.0
